ICU-535 Fixed some compiler warnings
X-SVN-Rev: 3602

+++ icu4c/source/test/intltest/numrgts.h
@@ -135,7 +135,7 @@ public:
             Formattable&            result, 
             ParsePosition&          /* parsePosition */) const
     {
-        result.setLong(0L);
+        result.setLong((int32_t)0);
     }
   
     virtual void parse(    const UnicodeString&    text,
